+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Fix `otherside` and `netherside` examples force-loading backup worlds
  - The Door action now sets the target location to the bottom block of the door
  - Hex colors now supported in 1.16 (as well as 1.17)
+ - Maybe support custom enchantment plugins (1.17 only for now)
 
 # 9.0